% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry_penetrance.R
\name{ancestry_penetrance}
\alias{ancestry_penetrance}
\title{Compute and Plot Ancestry-Stratified Max/Median Penetrance by ACMG Condition}
\usage{
ancestry_penetrance(ah_low, ah_high, dataset, range, position, af_type)
}
\description{
Compute and Plot Ancestry-Stratified Max/Median Penetrance by ACMG Condition
}
\examples{
ancestry_pen <- ancestry_penetrance(ah_low = 0.01, ah_high = 1, dataset = "gnomAD", 
range = 5, position = "Max", af_type = "gene")
}
