% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_1000g}
\alias{annotate_1000g}
\title{Annotate 1000 Genomes Variant VCF using ClinVar}
\usage{
annotate_1000g(clinvar, vcf)
}
\arguments{
\item{clinvar}{data.frame; clinvaR-processed vcf containing ClinVar data. 
Defaults to get_clinvar().}

\item{vcf}{data.frame; clinvaR-processed vcf containing 1000 genomes sequencing data. 
Defaults to importing data from all downloads: vcf <- import_file_1000g().}
}
\description{
This function takes a merged ClinVar-sequencing dataset and 
returns an ancestry-stratified plot of allele frequencies.
}
\examples{
annotate_1000g(genes = get_genes())
annotate_1000g(clinvar = get_clinvar(), vcf = import_file_1000g())
}
