% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_UCSC.R
\name{connect_UCSC}
\alias{connect_UCSC}
\title{Connect to UCSC hg19 Genome Database}
\usage{
connect_UCSC
}
\description{
This function makes a connection to the hg19 genome database and defines a query function.
}
