% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinvar.R
\name{download_clinvar}
\alias{download_clinvar}
\title{Import ClinVar VCF}
\usage{
download_clinvar()
download_clinvar(file)
}
\arguments{
\item{file}{character; specifies the path to the VCF.}
}
\description{
This function allows you to read a ClinVar VCF into a table 
and extract important information from the INFO section.
}
\details{
getclinvar() collects the latest VCF from 'ftp://ftp.ncbi.nlm.nih.gov/pub/clinvar/vcf_GRCh37/clinvar.vcf.gz'
Specifying a file path will collect the VCF from that path instead.
}
\examples{
clinvar <- sprintf("ClinVar_Reports/VCF/clinvar_\%s.vcf.gz", clinvar_date) \%>\% 
              download_clinvar(file)
clinvar <- clinvar[!duplicated(clinvar$VAR_ID),] # remove duplicates
}
