% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_CI.R
\name{freq_CI}
\alias{freq_CI}
\title{Compute One-Tailed Beta Distribution Confidence Interval from Observed Alleles}
\usage{
freq_CI(freq, n, cutoff)
}
\arguments{
\item{freq}{numeric; observed frequency}

\item{n}{numeric; number of observations, equal to the number of observed alleles 
(twice the number of observed individuals). This is hard coded using dataset FAQs for 
the aggregateCount() and aggregateCalc() functions}

\item{cutoff}{numeric; upper tailed confidence range. Default is 95% = 0.95.}
}
\description{
This function computes a one-Tailed Beta distribution confidence interval from observed alleles
}
\examples{
freq_CI(freq = 0.05, n = 2*2504, cutoff = 0.95)
}
