% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_over_time.R
\name{freq_over_time}
\alias{freq_over_time}
\title{Population frequencies over time}
\usage{
freq_over_time(vcf)
freq_over_time(genes)
}
\arguments{
\item{genes}{character; genes of interest.}

\item{vcf}{data.frame; clinvaR-processed vcf from import_file_1000g()}
}
\description{
This function computes the probability of having at least 1 variant  
in the gene list of interest (assuming independence) for a range of ClinVar files
}
\examples{
freq_over_time(get_genes('lmm_hcm'))
freq_over_time(import_file_1000g())
}
