% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_allele_freq.R
\name{get_allele_freq}
\alias{get_allele_freq}
\title{Compute Aggregate Allele Frequencies by ACMG Condition}
\usage{
get_allele_freq(input, ind, dataset, method)
}
\arguments{
\item{input}{data frame; merged ClinVar-sequencing dataset. Defaults to merge_clinvar_[dataset]().}

\item{ind}{logical; compute using independence assumption. 
FALSE is only permitted for 1000 Genomes data. Defaults to TRUE.}

\item{dataset}{character; choose from '1000 Genomes','ExAC', or 'gnomAD'. 
Not case-sensitive. Defaults to 'gnomAD'.}

\item{method}{character; choose from 'Gene','MIM','MedGen'. Not case-sensitive. Defaults to 'Gene'.}
}
\description{
Compute Aggregate Allele Frequencies by ACMG Condition
}
\examples{
freq_gnomad.calc.gene <- get_allele_freq()
freq_1000g.count.gene <- get_allele_freq(input = merged_1000g, ind = F, dataset = "1000G", method = "Gene")
freq_exac.calc.gene <- get_allele_freq(input = merged_exac, ind = T, dataset = "EXAC",method = "Gene")
}
