% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinvar.R
\name{get_clinvar}
\alias{get_clinvar}
\title{Import ClinVar VCF}
\usage{
get_clinvar()
get_clinvar(file)
}
\arguments{
\item{file}{Specifies the path to the VCF.}
}
\description{
This function allows you to read a ClinVar VCF 
into a table and extract important information
from the INFO section.
}
\details{
getclinvar() collects the latest VCF from 'ftp://ftp.ncbi.nlm.nih.gov/pub/clinvar/vcf_GRCh37/clinvar.vcf.gz'
Specifying a file path will collect the VCF from that path instead.
}
\examples{
clinvar_date <- system("ls ClinVar_Reports/VCF/clinvar*.vcf.gz", intern = T) \%>\% 
strsplit("clinvar_") \%>\% sapply("[[",2) \%>\% unlist \%>\% 
strsplit(".vcf.gz")  \%>\% sapply("[[",1) \%>\% unlist \%>\% as.Date() \%>\% max
# collects the most recent date in './ClinVar_Reports/VCF'
clinvar_file <- sprintf("ClinVar_Reports/VCF/clinvar_\%s.vcf.gz", clinvar_date)
# find the file name
clinvar <- get_clinvar(clinvar_file)
clinvar <- clinvar[!duplicated(clinvar$VAR_ID),] # remove duplicates
}
