% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_genes}
\alias{get_genes}
\title{Imports List of Genes from File}
\usage{
get_genes(file)
}
\arguments{
\item{file}{character; specified file location, or local file names from extdata/MacArthur_Gene_Lists/gene_lists/lists}
}
\description{
This function imports a list of genes from a specified file location, or from the MacArthur gene lists.
}
\examples{
get_genes('lmm_hcm.tsv')
}
