% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_penetrance.R
\name{get_penetrance}
\alias{get_penetrance}
\title{Compute and Plot Penetrance Ranges by ACMG Condition}
\usage{
get_penetrance(ah_low, ah_high, dataset)
}
\description{
Compute and Plot Penetrance Ranges by ACMG Condition
}
\examples{
pen_gnomad <- get_penetrance(ah_low = 0.01, ah_high = 1, 
dataset = "gnomAD", alleleFreq = freq_gnomad.calc.gene)
pen_1000g <-  get_penetrance(ah_low = 0.01, ah_high = 1, 
dataset = "1000 Genomes", freq_1000g.count.gene)
}
