% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinvar.R
\name{get_test_clinvar}
\alias{get_test_clinvar}
\title{Import Test ClinVar VCF}
\usage{
get_test_clinvar()
}
\description{
This function imports the binary ClinVar VCF from April 4, 2017. 
This skips the processing steps, allowing faster access for testing purposes
}
