% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlleleFreq.R
\name{getAlleleFreq}
\alias{getAlleleFreq}
\title{Compute Aggregate Allele Frequencies by ACMG Condition}
\usage{
getAlleleFreq(input, ind, dataset, method)
}
\description{
Compute Aggregate Allele Frequencies by ACMG Condition
}
\examples{
freq_1000g.count.gene <- getAlleleFreq(input = merged_1000g, ind = F, dataset = "1000G", method = "Gene")
freq_1000g.calc.gene <- getAlleleFreq(input = merged_1000g, ind = T, dataset = "1000G", method = "Gene")
freq_gnomad.calc.gene <- getAlleleFreq(input = merged_gnomad, ind = T, dataset = "GNOMAD", method = "Gene")
freq_exac.calc.gene <- getAlleleFreq(input = merged_exac, ind = T, dataset = "EXAC",method = "Gene")
}
