% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_read.R
\name{h_read}
\alias{h_read}
\title{Convert seconds to human-readable time units}
\usage{
h_read(timing)
}
\arguments{
\item{timing}{the number of elapsed seconds, returned by proc.time() - ptm}
}
\description{
This function converts a number of seconds created by proc.time() - ptm 
to days, hours, minutes, seconds.
}
