% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ACMG.R
\name{import_ACMG}
\alias{import_ACMG}
\title{Import ACMG Table from Data}
\usage{
import_ACMG()
}
\description{
This function imports Table 1 of the ACMG SF v2.0 paper
(http://www.nature.com/gim/journal/v19/n2/abs/gim2016190a.html)
and separates listed entries by the pipe delimiter.
}
\details{
Columns are: Phenotype, MIM_disorder, PMID_Gene_Reviews_entry, Typical_age_of_onset, 
Gene, MIM_gene, Inheritance, Variants_to_report

The file is stored in inst/extdata.
}
\keyword{ACMG}
