% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_file_exac.R
\name{import_file_exac}
\alias{import_file_exac}
\title{Imports downloaded VCF from ExAC or gnomAD}
\usage{
import_file_exac(gene, dataset)
}
\description{
This function imports a gene's VCF and extracts attributes into a 
specified format for other clinvaR analyses
}
