% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_clinvar.R
\name{merge_clinvar_1000g}
\alias{merge_clinvar_1000g}
\title{Merge ClinVar with 1000 Genomes}
\usage{
merge_clinvar_1000g(clinvar, ACMG.1000g)
}
\description{
This function takes a merged ClinVar-sequencing dataset and 
returns an ancestry-stratified plot of allele frequencies
}
\examples{
merge_clinvar_1000g(clinvar, ACMG.1000g)
}
