% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_clinvar.R
\name{merge_clinvar_exac}
\alias{merge_clinvar_exac}
\title{Merge ClinVar with ExAC}
\usage{
merge_clinvar_exac(clinvar, ACMG.1000g)
}
\description{
This function takes a merged ClinVar-sequencing dataset and 
returns an ancestry-stratified plot of allele frequencies
}
\examples{
merge_clinvar_exac(clinvar, ACMG.1000g)
}
