% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_over_time.R
\name{over_time}
\alias{over_time}
\title{ClinVar Over Time Analyses}
\usage{
over_time(vcf)
over_time(genes)
}
\arguments{
\item{vcf}{data.frame; clinvaR-processed vcf from import_file_1000g()}

\item{genes}{character; list of genes of interest.}
}
\description{
This function computes aggregate frequency, number of related positions, 
number of related assertions, and (if given case allele frequency and prevalence)
aggregate penetrance across all stored ClinVar files.
Returns a list of 2 or 3
}
\examples{
over_time(import_file_1000g())
}
