% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_over_time.R
\name{penetrance_over_time}
\alias{penetrance_over_time}
\title{Aggregated penetrance over time}
\usage{
penetrance_over_time(vcf)
}
\arguments{
\item{vcf}{data.frame; clinvaR-processed vcf from import_file_1000g()}
}
\description{
This function computes the penetrance of a positive finding
in the gene list of interest (assuming independence) 
for a range of ClinVar files. Prevalence and Case Frequency values must be given.
}
\examples{
penetrance_over_time(import_file_1000g())
}
