% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_over_time.R
\name{plot.freq_dates}
\alias{plot.freq_dates}
\title{Plot Aggregate Population Frequencies Over Time}
\usage{
plot(freqs)
}
\arguments{
\item{freqs}{the named frequencies output by freq_over_time().}
}
\description{
This function plots the aggregate population frequency generated by freq_over_time().
}
