% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_VCF.R
\name{plot.plain_vcf}
\alias{plot.plain_vcf}
\title{Compute and Plot Aggregate Allele Frequencies for 1000 Genomes}
\usage{
var_plot_1000g(vcf)
var_plot_1000g(vcf, fraction = FALSE)
}
\arguments{
\item{vcf}{plain_vcf data.frame; clinvaR-processed VCF from 1000 Genomes.}

\item{fraction}{logical; if TRUE, plots fraction with a finding. If FALSE, plots total counts.}
}
\description{
This function takes a plain ClinVar-sequencing dataset and 
returns an ancestry-stratified plot of counts or frequencies.
}
