% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over_time.R
\name{plot.significance_dates}
\alias{plot.significance_dates}
\title{Plot Significances Over Time}
\usage{
plot(significances)
}
\arguments{
\item{significances}{the named signifance output generated by over_time().}
}
\description{
This function plots the aggregate counts of variants
of each significances category, as generated by over_time().
}
