% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyprint.R
\name{prettyprint}
\alias{prettyprint}
\title{Barplot of ancestry-stratified 1000 Genomes allele frequencies}
\usage{
prettyprint(values, title = NULL, sd = TRUE, ylimit = NULL, xlabel = "Population", ylabel = NULL)
}
\description{
Barplot of ancestry-stratified values
}
\examples{
prettyprint(values, title = title, sd = F, ylimit = NULL, 
xlabel = "Population", ylabel = "Fraction with at least 1 non-reference site")
}
