% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ACMG.R
\name{scrape_ACMG}
\alias{scrape_ACMG}
\title{Scrape ACMG Table from ClinVar}
\usage{
scrape_ACMG()
}
\description{
This function scrapes the table from 'https://www.ncbi.nlm.nih.gov/clinvar/docs/acmg' and cleans it up. 
Columns are: Disease, MedGen, Gene, Variations_Link.
}
\keyword{ACMG}
