% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_over_time.R
\name{submissions_over_time}
\alias{submissions_over_time}
\title{Number of Submissions Over Time}
\usage{
submissions_over_time(genes)
}
\arguments{
\item{genes}{character; list of genes.}
}
\description{
This function computes the number of unique submissions for point mutations
(identified by chromosome, position, reference, alternate) in ClinVar
}
\examples{
submissions_over_time(get_genes('lmm_hcm.tsv'))
}
