% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinvar.R
\name{update_clinvar}
\alias{update_clinvar}
\title{Download Newer Versions of ClinVar}
\usage{
update_clinvar()
#@export
}
\description{
This function downloads, processes, and stores newer versions of ClinVar.
}
