% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_plot_1000g.R
\name{var_plot_1000g}
\alias{var_plot_1000g}
\title{Compute and Plot Aggregate Allele Frequencies for 1000 Genomes}
\usage{
var_plot_1000g(vcf)
var_plot_1000g(vcf, fraction = FALSE)
}
\arguments{
\item{vcf}{data.frame; clinvaR-processed VCF from 1000 Genomes. Defaults to importing from genes (if present).}

\item{fraction}{logical; if TRUE, plots fraction with a finding. If FALSE, plots total counts.}
}
\description{
This function takes a merged ClinVar-sequencing dataset and 
returns an ancestry-stratified plot of allele frequencies
}
