% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_plot_1000g.R
\name{var_plot_exac}
\alias{var_plot_exac}
\title{Compute and Plot Aggregate Allele Frequencies for ExAC}
\usage{
var_plot_exac(clinvar, ACMG.exac, pathogenic, frac)
}
\arguments{
\item{pathogenic}{logical; whether to limit frequency counts to pathogenic variants only.}

\item{frac}{logical; if TRUE, plots fraction with a finding. If FALSE, plots total counts.}
}
\description{
This function takes a merged ClinVar-sequencing dataset and 
returns an ancestry-stratified plot of allele frequencies
}
\examples{
var_plot_exac(clinvar, ACMG.exac, pathogenic = T, frac = T)
}
