% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_plot_1000g.R
\name{var_plot_gnomad}
\alias{var_plot_gnomad}
\title{Compute and Plot Aggregate Allele Frequencies for gnomAD}
\usage{
var_plot_gnomad(clinvar, ACMG.gnomad, pathogenic, frac)
}
\description{
This function takes a merged ClinVar-sequencing dataset and 
returns an ancestry-stratified plot of allele frequencies
}
\examples{
var_plot_gnomad(clinvar, ACMG.gnomad, pathogenic = T, frac = T)
}
