% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry_penetrance.R
\name{ancestry_penetrance}
\alias{ancestry_penetrance}
\title{Compute and Plot Ancestry-Stratified Max/Median Penetrance by ACMG Condition}
\usage{
ancestry_penetrance(ah_low, ah_high, dataset, position, alleleFreq)
}
\arguments{
\item{ah_low}{numeric; lower allelic heterogeneity bound.}

\item{ah_high}{numeric; hgher allelic heterogeneity bound.}

\item{dataset}{character; dataset name. Choose from 1000 Genomes, ExAC, or gnomAD (not case sensitive)}

\item{position}{character; max or Median. Defaults to Max.}

\item{alleleFreq}{data frame (31 x n); contains the columns 'AF_[dataset]_[population]' 
for 5 superpopulations and the general population. Typically an output from getAlleleFreq().}
}
\description{
This function uses prevalence and allele frequency data to compute point and range 
estimates for penetrance. These are used to generate three plots: 
(1) Radar Plot; (2) Faceted Bar Plot; (3) Heat Map.
The function returns the computed table of penetrance values (disease-conditions by superpopulations).
}
\examples{
ancestry_pen <- ancestry_penetrance(ah_low = 0.01, ah_high = 1, dataset = "gnomAD", 
position = "Max", alleleFreq = freq_gnomad.calc.gene)
}
