% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_1000g.R
\name{download_1000g}
\alias{download_1000g}
\title{Downloads VCF from 1000 Genomes using tabix}
\usage{
download_1000g(genes, download)
}
\arguments{
\item{genes}{character; name of a gene(s) indexed in refGene}

\item{download}{logical; indicates whether to download the file. 
FALSE still returns transcription coordinates and RefGene details from UCSC.}
}
\description{
This function downloads a gene's VCF from the 1000 Genomes FTP based on its transcription region
as listed in the UCSC Genome database.
}
\examples{
download_1000g('BRCA2', download = T)
download_output <- sapply(ACMG.panel, function(gene) download_1000g(gene, download = T)) \%>\% t
}
