% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_CI.R
\name{freq_CI}
\alias{freq_CI}
\title{Compute One-Tailed Beta Distribution Confidence Interval from Observed Alleles}
\usage{
freq_CI(freq, n, cutoff)
}
\description{
This function computes a one-Tailed Beta distribution confidence interval from observed alleles
}
\examples{
freq_CI(freq = 0.05, n = 2*sample_size, cutoff = 0.95)
}
