% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinvar.R
\name{get_clinvar}
\alias{get_clinvar}
\title{Get ClinVar Version from Date}
\usage{
get_clinvar()
get_clinvar(date)
}
\arguments{
\item{file}{character; input date in the format \%y-\%m-\%d. Ex: 2017-07-05.}
}
\description{
This function returns the ClinVar version from the date closest to the input date. 
If there is no input, it defaults to the current date.
}
