% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_VCF.R
\name{import_file_1000g}
\alias{import_file_1000g}
\title{Imports downloaded VCF from 1000 Genomes}
\usage{
import_file_1000g(gene, dataset, path)
}
\arguments{
\item{gene}{character; any of the pre-stored genes in extdata/1000G
Not case-sensitive. Defaults to 'gnomAD'.}

\item{path}{character; path to an importable VCF. Overrides gene input.}
}
\description{
This function imports a gene's VCF and extracts attributes from INFO 
into a specified format for other clinvaR analyses
}
\examples{
import_file_1000g('BRCA2')
ACMG.1000g <- NULL
for (gene in ACMG.panel) {
ACMG.1000g <- rbind(ACMG.1000g,import_file_1000g(gene))
}
#ACMG.1000g[duplicated(ACMG.1000g$VAR_ID),1:8] # Display and remove duplicates
ACMG.1000g <- ACMG.1000g[!duplicated(ACMG.1000g$VAR_ID),]
#@export
}
