% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_VCF.R
\name{import_file_exac}
\alias{import_file_exac}
\title{Imports downloaded VCF from ExAC or gnomAD}
\usage{
import_file_exac(gene, dataset)
}
\description{
This function imports a gene's VCF and extracts attributes into a 
specified format for other clinvaR analyses
}
\examples{
import_file_exac('BRCA2', 'gnomad')
ACMG.exac <- NULL; ACMG.gnomad <- NULL
for (gene in ACMG.panel) {
  print(sprintf("[\%d/\%d] \%s",which(gene==ACMG.panel),length(ACMG.panel),gene))
  ACMG.exac <- rbind(ACMG.exac,import_file_exac(gene, "exac"))
  ACMG.gnomad <- rbind(ACMG.gnomad,import_file_exac(gene, "gnomad"))
}
ACMG.exac <- ACMG.exac[!duplicated(ACMG.exac$VAR_ID),]
ACMG.gnomad <- ACMG.gnomad[!duplicated(ACMG.gnomad$VAR_ID),]
}
