% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_clinvar.R
\name{merge_clinvar_1000g}
\alias{merge_clinvar_1000g}
\title{Merge ClinVar with 1000 Genomes}
\usage{
merge_clinvar_1000g(clinvar, vcf)
}
\arguments{
\item{clinvar}{data.frame; clinvaR-processed vcf containing ClinVar data. 
Defaults to get_clinvar().}

\item{ACMG.1000g}{data.frame; clinvaR-processed vcf containing 1000 genomes sequencing data. 
Defaults to 'extdata/Supplementary_Files/ACMG_1000G.rds'.}
}
\description{
This function takes a merged ClinVar-sequencing dataset and 
returns an ancestry-stratified plot of allele frequencies.
}
\examples{
merge_clinvar_1000g()
merge_clinvar_1000g(clinvar = get_clinvar(), vcf = import_file_1000g())
}
