% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_clinvar.R
\name{merge_clinvar_exac}
\alias{merge_clinvar_exac}
\title{Merge ClinVar with ExAC}
\usage{
merge_clinvar_exac(clinvar, ACMG.exac)
}
\arguments{
\item{clinvar}{clinvaR-processed vcf containing ClinVar data. 
Defaults to get_test_clinvar().}

\item{ACMG.exac}{clinvaR-processed vcf containing ExAC sequencing data. 
Defaults to 'extdata/Supplementary_Files/ACMG_EXAC.rds'.}
}
\description{
This function takes a merged ClinVar-sequencing dataset and 
returns an ancestry-stratified plot of allele frequencies
}
\examples{
merge_clinvar_exac()
merge_clinvar_exac(clinvar, ACMG.exac)
}
