% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_clinvar.R
\name{merge_clinvar_gnomad}
\alias{merge_clinvar_gnomad}
\title{Merge ClinVar with gnomAD}
\usage{
merge_clinvar_gnomad(clinvar, ACMG.gnomad)
}
\arguments{
\item{clinvar}{clinvaR-processed vcf containing ClinVar data. 
Defaults to get_test_clinvar().}

\item{ACMG.gnomAD}{clinvaR-processed vcf containing gnomAD sequencing data. 
Defaults to 'extdata/Supplementary_Files/ACMG_GNOMAD.rds'.}
}
\description{
This function takes a merged ClinVar-sequencing dataset and 
returns an ancestry-stratified plot of allele frequencies
}
\examples{
merge_clinvar_gnomad()
merge_clinvar_gnomad(clinvar, ACMG.gnomad)
}
