% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_over_time.R
\name{plot.penetrance_dates}
\alias{plot.penetrance_dates}
\title{Plot Aggregated Penetrance Over Time}
\usage{
plot(penetrance)
}
\arguments{
\item{penetrance}{the named penetrance output by penetrance_over_time().}
}
\description{
This function plots the aggregated penetrance values 
generated by penetrance_over_time().
}
