% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_over_time.R
\name{plot.submission_dates}
\alias{plot.submission_dates}
\title{Plot Submissions Over Time}
\usage{
plot(submissions)
}
\arguments{
\item{submissions}{the named frequencies output generated by over_time().}
}
\description{
This function plots the aggregate counts of positions
and assertions generated by over_time().
}
