% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_ACMG.R
\name{scrape_ACMG}
\alias{scrape_ACMG}
\title{Scrape ACMG Table from ClinVar}
\usage{
scrape_ACMG(scrape)
}
\description{
This function scrapes the table from 'https://www.ncbi.nlm.nih.gov/clinvar/docs/acmg' and cleans it up.
}
\keyword{ACMG}
